Detect and fix printf signedness errors. Fixes #1086, #1087.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10846 3db68df2-c116-0410-a063-a993310a9797

+++ source/maths/tests/test_MD5.h
@@ -26,7 +26,7 @@ public:
 	{
 		char digeststr[MD5::DIGESTSIZE*2+1];
 		for (size_t i = 0; i < MD5::DIGESTSIZE; ++i)
-			sprintf_s(digeststr+2*i, 3, "%02x", digest[i]);
+			sprintf_s(digeststr+2*i, 3, "%02x", (unsigned int)digest[i]);
 		return digeststr;
 	}
